Moves is_lvalue to expr_util.c
This change exposes is_lvalue for use outside of
goto-programs/builtin_functions.cpp.
